Add support for arrays and matrices in scanner/parser
Implemented support for arrays/matrices in the scanner,
and modifies boolean operators to match our syntax. Also
did some renaming (like with the last commit) to make things
more consistent.

Support for arrays/matrices in the parser is incomplete; it
supports parsing the array/matrix types, but not for parsing
the literals just yet (want to hold off until we decide on
a way to differentiate between array and matrix literals).