Merge LANGLE/LT and RANGLE/RT
It might be simpler to just treat LANGLE/LT and RANGLE/RT
as the same token, and leave balancing to the parser,
rather than to the scanner. The code is now simpler, since
there's no need for the "stateful" recursive rules in the
scanner, and the grammar rule in the parser handles the
balancing now. On the other hand, we now also use LANGLE/RANGLE
for Less/Greater, which is not ideal, but maybe better overall?