Support initialization for decls and decls as statements
Variable declarations can now also be initialized with an
expression, rather than forcing declarations and assignments
to be separate.

Also, variable declarations are included as part of statements,
meaning that interweaving of declarations are allowed in the
program, rather than forcing them to be declared at the top of
any function.