Add support for index/slicing operators
Changed double operators to use commas for readability and
grammar reasons.

Also restructured grammar to enforce precedence of unary
and postfix operators (LALR parser leads to shift/reduce
conflict otherwise, even with the precedence tokens set).

As a result, slightly more work is passed to SAST, as we
need to validate that for assignments, the assignee is
actually valid (i.e. ID or index expression). Shouldn't be
too hard.