Support scope for blocks/for loops + function return check
Blocks are no longer flattened, and provide their own local
scope. For loop initializers that are declarations are now
implemented as if the entire for loop was in its own block,
with a preceding variable declaration, to avoid duplicate
declarations from outside the loop (i.e. effectively the
declaration is considered as part of the for loop's scope).

Also added checks to see if a function with a non-void return
type actually contains return statements.