Reimplement for initializer
Instead of having a special initializer type,
whenever we see a declaration initializer in a
for loop, we just replace it with a block that has
a declaration followed by a for loop (with an
empty initializer).

Also cleaned up some of the formatting, and fixed
the scanner tests (there was a missing option that
passed test cases through the entire stack, rather
than just the Ast pass).