Implement arrays, remove malloc calls, optimize semant
Implemented arrays in Neo. Added a preliminary print_array
function that isn't fully supported (TODO: implement way
in codegen for generating the print_element functions based
on array type).

Rewrote the way that we do array/matrix creation in codegen
so that there's no need for mallocs in native.c; the only time
we'll need to malloc/free is whenever one of the Neo functions
needs to return a matrix or array; we'll have to figure out a
way of doing this later.

Also changed some of the organization in semant so that only
one pass is needed; basically we're carrying the current scope
along with us as we pass through the code and modifying the scope
as we go along, as opposed to building up the entire scope first
before passing it.