Implement function pointers + function types
Proper support for passing around function pointers works now;
had to restructure a bit of semant/codegen to make this work.

Also did some reformatting of semicolons for consistency.

Deleted transpose implementation/tests, as we'll probably just
implement that in our standard library (one for each type);
we're planning on doing that for matrix functions anyways for
convenience, and removing this already-implemented one keeps it
consistent.

Maybe at some point we can do templated functions to make this
easier, but that'll be a stretch.