Add 'with' keyword, allow arbitrary expressions for protest
Changed protest syntax to be `protest [exc] with [val]`, as
it'd be nice to have `[exc]` be an arbitrary expression (in
other words, we can pass exception types around like variables,
though this is the only use case for it). The previous syntax
would result in a shift/reduce conflict in the parser, so the
addition of the `with` keyword disambiguates this situation.