Implement support for null expression runtime checks
Support for runtime checks is done in the form of a check
function, which checks if a condition is true. If false,
it prints to stderr and exits with EXIT_FAILURE.

Support for checking null values is now done with an appropriate
test for empty arrays (since they can be filled with null values,
if the element type is not a basic data type).