Reimplement native container ops to internally malloc results
Instead of codegen providing a malloc'd result for matrix ops
(such as slicing, binops) or array ops (slicing), since these
all make calls to native.c and we already have functionality to
malloc an entire matrix, codegen has been modified to simply
make calls to these functions, and these functions now figure out
what arrays/matrices to malloc, and return these.

This used to be necessary when arrays/matrices are stored on
the stack, so we didn't want to lose data, but now that they're
on the heap, it doesn't matter. We just need to make sure to
properly free them (if the user wants) later.

Also did some formatting stuff with pointer types; I think it's
better to write int* i rather than int *i, as it's more clear,
but this is a personal preference (also stems from C++ devs).