Improve naming for internal functions
Replaced the term "native" for "internal", which is more accurate;
native probably makes less sense for what these functions are, which
is simply just the compiler's internal implementation for different
features.

As a result, also renamed `native.c` and `native.h` to `libneoc.c`
and `libneoc.h`, to indicate that these files are the internal library
for the compiler, which is `neoc` (similar naming convention to how
Rust names their internal library, for example).