Add optimization pass to clean dead code and collect uses
We now add a second pass over the SAST, which collects any
function uses (both user-declared and internal). We then perform
a graph search, and eliminate any functions that won't be executed;
this cuts down on generated LLVM.

This is also cleaner, as we can then use this pass to gather any
internal/helper function uses, instead of making our semant pass
even messier. This also finally allows us to get rid of the
`translation_environment` record, and we can simply pass the
scope around during the semant instead of extra baggage.